requirements.txt
Jinja2==3.1.2
numpy==1.22.0
pandas==1.4.2
pdfkit==1.0.0
python_dateutil==2.8.2
